First pass at fixing race conditions when initially connecting to a Forge server.
When a user first connects to Bungee, once login has completed, connection to the Server starts. However, with Forge, we start the handshake at the same time as we start the server connector. In some cases, the handshake completes from Bungee <-> Client before Bungee <-> Sever has even seen a hello. However, in the majority of cases, the server handshakes first, and doesn't think the Client is a forge client, disconnecting them for the reason that they are not modded, even though they are.

This commit introduces a delay if a Server Hello is encountered. If so, the packet is stored, and handled again if the client then completes a FML handshake. If not after a set time (currently 5 seconds), the user is then disconnected.

This commit should not be considered final (there might be some cleaning up to do), but should fix the issue.