Add an API call on the ProxiedPlayer, getModList, to get an immutable map of user's mods along with their versions if a FML handshake has been completed. If a FML handshake has not been completed, returns an empty map.
The major use case for this is for a potential Bungee plugin that can disconnect a user when it detects a mod that is on a blacklist/not on a whitelist - but there are probably other uses for it too. More information can be found in the JavaDocs for net.md_5.bungee.api.connection.ProxiedPlayer.getModList.

Conflicts:
	api/src/main/java/net/md_5/bungee/api/connection/ProxiedPlayer.java
	proxy/src/main/java/net/md_5/bungee/UserConnection.java